Major refactoring dell'API di questionnaireManager. Aggiunta la classe questionnaireUtility. Spostato il codice di answerManager e rimossa la classe (non previsa dal Class Diagram ed inutile). Spostato il conteggio delle risposte di una certa domanda dal frontend al backend.

+++ api/questionnaireManager.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
 const mailProvider = require('./mailProvider')
+const questionnaireUtility = require('./questionnaireUtility')
 
 const AuthenticatedUser = require('../models/authenticatedUser');
-const QuestionOption = require('../enums/questionOption.cjs')
 const QuestionType = require('../enums/questionType.cjs')
 const Question = require('../models/question');
 const Answer = require('../models/answer');
@@ -18,30 +18,14 @@ const QUESTIONNAIRE_SUBMISSION_COOLDOWN_MIN = 1_440 //24h
 
 const API_V = process.env.API_VERSION;
 
-//permette di aggiungere argomenti dinamicamente ad una stringa.
-String.prototype.format = function() {
-    let formatted = this;
-    for (let i = 0; i < arguments.length; i++) {
-        let regexp = new RegExp('\\{' + i + '\\}', 'gi');
-        formatted = formatted.replace(regexp, arguments[i]);
-    }
-    return formatted;
-};
-
 router.get("/", async (req, res, next) => {
     if (req.query.type == "questionnaire"){
         let start = new Date(Date.now());
         if (LOG_MODE >= 2) console.log("[Get questionnaire] request started at "+ (start.getSeconds() * 1000 + start.getMilliseconds()))
         let questionList = await Question.find({}),
-        answeredQuestionList = await Answer.find({submitterId: req["loggedUser"].id});
+        userAnswerList = await Answer.find({submitterId: req["loggedUser"].id});
 
-        questionList = questionList.filter(question => {
-            let submitted = false;
-            answeredQuestionList.forEach(answered => {
-                if (String(answered.questionId) == String(question._id)) submitted = true;
-            });
-            if (!submitted) return question;
-        })
+        questionList = questionnaireUtility.getNotAnsweredQuestions(userAnswerList, questionList);
 
         if (LOG_MODE >= 2){
             let finish = new Date(Date.now());
@@ -49,86 +33,34 @@ router.get("/", async (req, res, next) => {
             console.warn("Query speed: "+(finish - start)+"ms");
         }
 
-        if (questionList.length >= QUESTIONNAIRE_MIN_QUESTION_COUNT){
-            let iterationCount = 0
-            while (questionList.length > QUESTIONNAIRE_MIN_QUESTION_COUNT && iterationCount < 999){
-                questionList.splice(Math.floor(Math.random() * questionList.length), 1)
-                iterationCount++;
-            }
-            if (iterationCount >= 999) 
-                return res.status(500).json({error: true, message: "Something went wrong while choosing random questions! More than 999 iterations!"});
-            else{
-                questionList = questionList.map(element => {
-                    return {
-                        self: API_V + '/questionnaires/' + element._id,
-                        question: element.question,
-                        questionType: element.questionType,
-                        options: element.options,
-                    };
-                });
-                return res.status(200).json(questionList);
-            }
+        questionList = questionnaireUtility.getRandomQuestions(questionList, QUESTIONNAIRE_MIN_QUESTION_COUNT);
+
+        if (questionList != null){ // getRandomQuestions restituisce null se  numberOfQuestions > questionList.length
+            questionList = questionList.map(element => {
+                return {
+                    self: API_V + '/questionnaires/' + element._id,
+                    question: element.question,
+                    questionType: element.questionType,
+                    options: element.options,
+                };
+            });
+            return res.status(200).json(questionList);
         }else
             return res.status(400).json({error: true, message: "You have compiled all the available questions. Please come back later!"});
     }else if (req.loggedUser.administrator == true || TEST_MODE){
         if (req.query.type == "answer" && (req.query.method == "email" || req.query.method == "browser") ){
             if (LOG_MODE >= 1) console.log("Get all submitted answers request!")
 
-            let answerList = await Answer.find({});
+            let answerList = await Answer.find({}), questionList = await Question.find({});
+            questionList = questionnaireUtility.getQuestionsWithAnswerData(questionList, answerList);
 
-            if (req.query.method == "browser") res.status(200).json(answerList);
+            if (LOG_MODE >= 3) console.log(questionList);
+
+            if (req.query.method == "browser")
+                res.status(200).json(questionList);
             else{
-                let questionList = await Question.find({})
-                questionList = questionList.map(element => {
-                    let newOptions = element.options, newAnswers = [];
-                    if (element.questionType == QuestionType.DICHOTOMOUS) newOptions = QuestionOption.DICHOTOMOUS
-                    else if (element.questionType == QuestionType.RATING_SCALE) newOptions = QuestionOption.RATING_SCALE
-                    if (newOptions) newOptions.forEach(option => newAnswers.push(0))
-                    return {
-                        _id: element._id,
-                        question: element.question,
-                        questionType: element.questionType,
-                        options: newOptions,
-                        answers: newAnswers,
-                        openAnswers: []
-                    }
-                });
-
-                answerList.forEach(answer => {
-                    let question = questionList.find(element => String(element._id) == String(answer.questionId));
-                    if (question.questionType != QuestionType.OPEN_ENDED){
-                        let optionIndex = question.options.indexOf(answer.answer)
-                        question.answers[optionIndex]++;
-                    }else
-                        question.openAnswers.push(answer.answer)
-                })
-
-                if (LOG_MODE >= 3) console.log(questionList);
-
                 let mailText = 'The following data shows all the questionnaire\'s answers organized by questions' + '\n';
-                let tableLine = "{0}=========================================================================================={1}",
-                    tableHeader = " {0}"+'\n'+
-                                  tableLine.format(">","<")+'\n';
-
-                let questionNumber = 1;
-                questionList.forEach(question => {
-                    mailText += "\n"
-                    mailText += tableHeader.format(questionNumber + ") " + question.question);
-
-                    if (question.questionType != QuestionType.OPEN_ENDED){
-                        let i = 0;
-                        question.answers.forEach(count => {
-                            mailText += " "+question.options[i] +": " + count+"\n"
-                            i++;
-                        });
-                    }else {
-                        question.openAnswers.forEach(answer => {
-                            mailText += " "+answer+"\n"
-                        });
-                    }
-                    mailText += tableLine.format(">","<")+"\n"
-                    questionNumber++;
-                })
+                mailText += questionnaireUtility.formatQuestionsForMail(questionList);
 
                 let mailOptions = {
                     subject: '[GreenMap] Questionnaire\'s answers',
@@ -267,6 +199,7 @@ router.post("/",  async (req, res) => {
         return res.status(400).json({error: true, message: "NON E' STATA PASSATA UNA QUESTION AL METODO!"});
 });
 
+//Aggiorna una question con del nuovo testo, tipo di domanda oppure opzioni fornite dall'amministratore
 router.put("/:questionId",  async (req, res) => {
     let submittedQuestion = req.body.question;
     if (submittedQuestion){
@@ -295,20 +228,39 @@ router.put("/:questionId",  async (req, res) => {
         return res.status(400).json({error: true, message: "NON E' STATA PASSATA UNA QUESTION AL METODO!"});
 });
 
+//Cancella una question e tutte le sue risposte a cascata
 router.delete("/:questionId",  async (req, res) => {
     if (req.loggedUser.administrator == true || TEST_MODE){
+        if (req.query.type == "question"){
+            try {
+                await Question.deleteOne({ _id: req.params.questionId })
+                await Answer.deleteMany({ questionId: req.params.questionId }) //cancella le risposte a cascata
+            } catch(err) {
+                return res.status(500).json({error: true, message: "An error occurred while deleting the question ",err});
+            }
 
-        try {
-            await Question.deleteOne({ _id: req.params.questionId })
-            await Answer.deleteMany({ questionId: req.params.questionId }) //cancella le risposte a cascata
-        } catch(err) {
-            return res.status(500).json({error: true, message: "An error occurred while deleting the question ",err});
-        }
-
-        if (LOG_MODE >= 1) console.log('Question successfully removed!');
-        res.status(204).send();
+            if (LOG_MODE >= 1) console.log('Question successfully removed!');
+            res.status(204).send();
+        }else if (req.query.type == "userAnswers"){
+            next();
+        }else
+            return res.status(400).json({error: true, message: "NON E' STATO PASSATO UN QUERY PARAMETER PREVISTO ALLA FUNZIONE!"});
     }else
         return res.status(401).json({error: true, message: 'Requesting user is not an administrator!'});
+});
+
+//[lavoro di raffaele, commit: d1f1ba8d7c44647c32c0218da98a90f4a129ff01]
+//Cancella tutte le risposte di un certo utente a cascata 
+router.delete("/:issuerId", async (req, res) => {
+    // SOPRA C'E' LA NEXT QUINDI NON SERVE IL CONTROLLO DELLE CREDENZIALI QUI!
+    if (LOG_MODE >= 1) console.log("Delete all answers of user " + req.params.issuerId);
+
+    try {
+        await Answer.deleteMany({ userId: req.params.issuerId })
+    }catch(err){
+        return res.status(400).json({error: true, message: "ID not found."});
+    }
+    return res.status(204).send();
 });
 
 router.delete("/",  async (req, res) => {
